Add SPDX headers and copyright notices at the beginning of files.
In downstream projects, files may be copied and moved around. This makes
it difficult to track the copyrights and licenses without having them
stored in each file.

I started with .reuse/dep5 but I've also noticed some things:
- for Files: *, there is no Copyright: entry which is required by the
  Debian machine-readable copyright format (and debian packaging is my
  ultimate goal)
- for Files: *, License: is not the same as in the LICENSE file (no AND
  MIT clause)

I've eyed making these changes in a way that would permit creating
.reuse/dep5 again in an automated fashion. This isn't possible due to
the differences I outline here and due to the tools available but that
was a goal (and explains why I changed */Makefile too).

While license was fine, I had an issue with copyrights which are not
indicated in the project but are required by Debian. I've used
"Copyright: TBD" in headers but I'd like to fix that obviously.

Moreover, I've chosen to use the most recent copyright/license/* header
for ref/{api.h, PQCgenKAT_sign.c,rng.c,rng.h} based on the files at
https://csrc.nist.gov/Projects/pqc-dig-sig/standardization/example-files
because "All rights reserved." is not a valid license. This had the
effect of changing Copyright: too but I think it is more meaningful now.
The actual code for these files has changed too but upon review, the
changes are all very minor.